Add mnemonic vaildation
